follow AS suggestions in com.ubergeek42.WeechatAndroid and refactor code
