+++ app/src/main/java/com/ubergeek42/WeechatAndroid/service/P.java
@@ -68,7 +68,6 @@ public class P implements SharedPreferences.OnSharedPreferenceChangeListener{
         loadUIPreferences();
         p.registerOnSharedPreferenceChangeListener(instance);
         _4dp = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, 4, context.getResources().getDisplayMetrics());
-        _50dp = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, 50, context.getResources().getDisplayMetrics());
         _200dp = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, 200, context.getResources().getDisplayMetrics());
         calculateWeaselWidth();
     }
@@ -98,7 +97,6 @@ public class P implements SharedPreferences.OnSharedPreferenceChangeListener{
     ///////////////////////////////////////////////////////////////////////////////////////////// ui
 
     public static float _4dp;
-    public static float _50dp;
     public static float _200dp;
 
     public static boolean sortBuffers;
@@ -147,7 +145,7 @@ public class P implements SharedPreferences.OnSharedPreferenceChangeListener{
         // buffer-wide preferences
         filterLines = p.getBoolean(PREF_FILTER_LINES, PREF_FILTER_LINES_D);
         autoHideActionbar = p.getBoolean(PREF_AUTO_HIDE_ACTIONBAR, PREF_AUTO_HIDE_ACTIONBAR_D);
-        maxWidth = Integer.parseInt(p.getString(PREF_MAX_WIDTH, PREF_MAX_WIDTH_D));
+        maxWidth = Integer.parseInt(getString(PREF_MAX_WIDTH, PREF_MAX_WIDTH_D));
         encloseNick = p.getBoolean(PREF_ENCLOSE_NICK, PREF_ENCLOSE_NICK_D);
         dimDownNonHumanLines = p.getBoolean(PREF_DIM_DOWN, PREF_DIM_DOWN_D);
         setTimestampFormat();
@@ -197,24 +195,24 @@ public class P implements SharedPreferences.OnSharedPreferenceChangeListener{
     @MainThread public static void loadConnectionPreferences() {
         host = p.getString(PREF_HOST, PREF_HOST_D);
         pass = p.getString(PREF_PASSWORD, PREF_PASSWORD_D);
-        port = Integer.parseInt(p.getString(PREF_PORT, PREF_PORT_D));
+        port = Integer.parseInt(getString(PREF_PORT, PREF_PORT_D));
         wsPath = p.getString(PREF_WS_PATH, PREF_WS_PATH_D);
 
         connectionType = p.getString(PREF_CONNECTION_TYPE, PREF_CONNECTION_TYPE_D);
         sshHost = p.getString(PREF_SSH_HOST, PREF_SSH_HOST_D);
-        sshPort = Integer.valueOf(p.getString(PREF_SSH_PORT, PREF_SSH_PORT_D));
+        sshPort = Integer.valueOf(getString(PREF_SSH_PORT, PREF_SSH_PORT_D));
         sshUser = p.getString(PREF_SSH_USER, PREF_SSH_USER_D);
         sshPass = p.getString(PREF_SSH_PASS, PREF_SSH_PASS_D);
         sshKey = FilePreference.getData(p.getString(PREF_SSH_KEY, PREF_SSH_KEY_D));
         sshKnownHosts = FilePreference.getData(p.getString(PREF_SSH_KNOWN_HOSTS, PREF_SSH_KNOWN_HOSTS_D));
 
-        lineIncrement = Integer.parseInt(p.getString(PREF_LINE_INCREMENT, PREF_LINE_INCREMENT_D));
+        lineIncrement = Integer.parseInt(getString(PREF_LINE_INCREMENT, PREF_LINE_INCREMENT_D));
         reconnect = p.getBoolean(PREF_RECONNECT, PREF_RECONNECT_D);
         optimizeTraffic = p.getBoolean(PREF_OPTIMIZE_TRAFFIC, PREF_OPTIMIZE_TRAFFIC_D);
 
         pingEnabled = p.getBoolean(PREF_PING_ENABLED, PREF_PING_ENABLED_D);
-        pingIdleTime = Integer.parseInt(p.getString(PREF_PING_IDLE, PREF_PING_IDLE_D)) * 1000;
-        pingTimeout = Integer.parseInt(p.getString(PREF_PING_TIMEOUT, PREF_PING_TIMEOUT_D)) * 1000;
+        pingIdleTime = Integer.parseInt(getString(PREF_PING_IDLE, PREF_PING_IDLE_D)) * 1000;
+        pingTimeout = Integer.parseInt(getString(PREF_PING_TIMEOUT, PREF_PING_TIMEOUT_D)) * 1000;
 
         if (Utils.isAnyOf(connectionType, PREF_TYPE_SSL, PREF_TYPE_WEBSOCKET_SSL)) {
             sslSocketFactory = SSLHandler.getInstance(context).getSSLSocketFactory();
@@ -255,7 +253,7 @@ public class P implements SharedPreferences.OnSharedPreferenceChangeListener{
                 BufferList.onGlobalPreferencesChanged(true);
                 break;
             case PREF_MAX_WIDTH:
-                maxWidth = Integer.parseInt(p.getString(key, PREF_MAX_WIDTH_D));
+                maxWidth = Integer.parseInt(getString(key, PREF_MAX_WIDTH_D));
                 BufferList.onGlobalPreferencesChanged(false);
                 break;
             case PREF_ENCLOSE_NICK:
@@ -318,7 +316,7 @@ public class P implements SharedPreferences.OnSharedPreferenceChangeListener{
     }
 
     @MainThread private static void setAlignment() {
-        String alignment = p.getString(PREF_PREFIX_ALIGN, PREF_PREFIX_ALIGN_D);
+        String alignment = getString(PREF_PREFIX_ALIGN, PREF_PREFIX_ALIGN_D);
         switch (alignment) {
             case "right":     align = Color.ALIGN_RIGHT; break;
             case "left":      align = Color.ALIGN_LEFT; break;
@@ -328,7 +326,7 @@ public class P implements SharedPreferences.OnSharedPreferenceChangeListener{
     }
 
     @MainThread private static void setTextSizeColorAndLetterWidth() {
-        textSize = Float.parseFloat(p.getString(PREF_TEXT_SIZE, PREF_TEXT_SIZE_D));
+        textSize = Float.parseFloat(getString(PREF_TEXT_SIZE, PREF_TEXT_SIZE_D));
         String bufferFont = p.getString(PREF_BUFFER_FONT, PREF_BUFFER_FONT_D);
 
         typeface = Typeface.MONOSPACE;
@@ -374,7 +372,6 @@ public class P implements SharedPreferences.OnSharedPreferenceChangeListener{
     static void setServiceAlive(boolean alive) {
         p.edit().putBoolean(ALIVE, alive).apply();
     }
-
 
     ////////////////////////////////////////////////////////////////////////////////////////////////
     //////////////////////////////////////////////////////////////////////////////////////////////// save/restore
@@ -474,5 +471,13 @@ public class P implements SharedPreferences.OnSharedPreferenceChangeListener{
         sentMessages.add(Utils.cut(line, 2000));
         if (sentMessages.size() > 40)
             sentMessages.pop();
+    }
+
+    ////////////////////////////////////////////////////////////////////////////////////////////////
+
+    // used to make warnings go away
+    @SuppressWarnings("ConstantConditions")
+    private static String getString(String key, String defValue) {
+        return p.getString(key, defValue);
     }
 }
