+++ app/src/main/java/com/ubergeek42/WeechatAndroid/utils/LinearLayoutManagerFix.java
@@ -1,7 +1,5 @@
-/**
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- */
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
 
 package com.ubergeek42.WeechatAndroid.utils;
 
@@ -9,6 +7,7 @@ import android.content.Context;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import android.util.AttributeSet;
 
+@SuppressWarnings("unused")
 class LinearLayoutManagerFix extends LinearLayoutManager {
 
     public LinearLayoutManagerFix(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
@@ -19,7 +18,7 @@ class LinearLayoutManagerFix extends LinearLayoutManager {
         super(context);
     }
 
-    public LinearLayoutManagerFix(Context context, int orientation, boolean reverseLayout) {
+    LinearLayoutManagerFix(Context context, int orientation, boolean reverseLayout) {
         super(context, orientation, reverseLayout);
     }
