+++ app/src/main/java/com/ubergeek42/WeechatAndroid/utils/Linkify.java
@@ -44,7 +44,7 @@ public class Linkify {
     // also make sure we don't crash if nothing can handle our intent
     private static class URLSpan2 extends URLSpan {
 
-        public URLSpan2(@NonNull String url) {super(url);}
+        URLSpan2(@NonNull String url) {super(url);}
 
         @Override public void updateDrawState(@NonNull TextPaint ds) {
             ds.setUnderlineText(true);
@@ -116,7 +116,7 @@ public class Linkify {
         "(?=" +
               // some possible punctuation
               // AND space or end of string
-              "[\\])>,.!?:\"”]*" +
+              "[])>,.!?:\"”]*" +
               "(?:\\s|$)" +
         ")"
         , Pattern.CASE_INSENSITIVE | Pattern.COMMENTS);
